+++ modules/titiler/iam.tf
@@ -1,119 +1,44 @@
-resource "aws_iam_role" "titiler_lambda_role" {
-  name_prefix = "titiler-${var.environment}-${data.aws_region.current.region}-lambdaRole"
-
-  assume_role_policy = <<EOF
-{
-  "Version": "2012-10-17",
-  "Statement": [
-    {
-      "Action": "sts:AssumeRole",
-      "Principal": {
-        "Service": "lambda.amazonaws.com"
-      },
-      "Effect": "Allow"
+data "aws_iam_policy_document" "titiler_lambda_assume_role" {
+  statement {
+    effect  = "Allow"
+    actions = ["sts:AssumeRole"]
+    principals {
+      type        = "Service"
+      identifiers = ["lambda.amazonaws.com"]
     }
-  ]
-}
-EOF
-
+  }
 }
 
-resource "aws_iam_policy" "titiler_lambda_policy" {
-  name_prefix = "titiler-${var.environment}-${data.aws_region.current.region}-lambdaPolicy"
-
-  policy = <<EOF
-{
-    "Version": "2012-10-17",
-    "Statement": [
-        {
-            "Action": [
-                "logs:CreateLogStream",
-                "logs:CreateLogGroup"
-            ],
-            "Resource": [
-                "arn:aws:logs:${data.aws_region.current.region}:${data.aws_caller_identity.current.account_id}:log-group:/aws/lambda/titiler*:*"
-            ],
-            "Effect": "Allow"
-        },
-        {
-            "Action": [
-                "s3:Get*",
-                "s3:List*"
-            ],
-            "Resource": "*",
-            "Effect": "Allow"
-        },
-        {
-            "Action": [
-                "logs:PutLogEvents"
-            ],
-            "Resource": [
-                "arn:aws:logs:${data.aws_region.current.region}:${data.aws_caller_identity.current.account_id}:log-group:/aws/lambda/titiler*:*:*"
-            ],
-            "Effect": "Allow"
-        }
-    ]
-}
-EOF
-
+resource "aws_iam_role" "titiler-lambda-role" {
+  name_prefix        = "titiler-lambdaRole"
+  assume_role_policy = data.aws_iam_policy_document.titiler_lambda_assume_role.json
 }
 
-resource "aws_iam_role_policy_attachment" "titiler_lambda_vpc_policy_attachment" {
-  role       = aws_iam_role.titiler_lambda_role.name
+data "aws_iam_policy_document" "titiler_lambda_policy" {
+  dynamic "statement" {
+    for_each = length(var.authorized_s3_arns) > 0 ? [1] : []
+    content {
+      sid       = "AuthorizedS3Access"
+      effect    = "Allow"
+      actions   = ["s3:GetObject"]
+      resources = var.authorized_s3_arns
+    }
+  }
+}
+
+resource "aws_iam_role_policy" "titiler-lambda-inline-policy" {
+  count  = length(var.authorized_s3_arns) > 0 ? 1 : 0
+  name   = "titiler-lambda-inline-policy"
+  role   = aws_iam_role.titiler-lambda-role.id
+  policy = data.aws_iam_policy_document.titiler_lambda_policy.json
+}
+
+resource "aws_iam_role_policy_attachment" "lambda_basic_execution" {
+  role       = aws_iam_role.titiler-lambda-role.name
+  policy_arn = "arn:aws:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole"
+}
+
+resource "aws_iam_role_policy_attachment" "lambda_vpc_access" {
+  role       = aws_iam_role.titiler-lambda-role.name
   policy_arn = "arn:aws:iam::aws:policy/service-role/AWSLambdaVPCAccessExecutionRole"
-}
-
-resource "aws_iam_role_policy_attachment" "titiler_lambda_base_policy" {
-  role       = aws_iam_role.titiler_lambda_role.name
-  policy_arn = aws_iam_policy.titiler_lambda_policy.arn
-}
-
-resource "aws_iam_role" "titiler_gw_role" {
-  name_prefix = "titiler-${var.environment}-${data.aws_region.current.region}-apigwRole"
-
-  assume_role_policy = <<EOF
-{
-  "Version": "2012-10-17",
-  "Statement": [
-    {
-      "Action": "sts:AssumeRole",
-      "Principal": {
-        "Service": "apigateway.amazonaws.com"
-      },
-      "Effect": "Allow"
-    }
-  ]
-}
-EOF
-}
-
-resource "aws_iam_policy" "titiler_gw_policy" {
-  name_prefix = "titiler-${var.environment}-${data.aws_region.current.region}-apigwPolicy"
-
-  policy = <<EOF
-{
-    "Version": "2012-10-17",
-    "Statement": [
-        {
-            "Action": [
-              "logs:CreateLogGroup",
-              "logs:CreateLogStream",
-              "logs:DescribeLogGroups",
-              "logs:DescribeLogStreams",
-              "logs:PutLogEvents",
-              "logs:GetLogEvents",
-              "logs:FilterLogEvents"
-            ],
-            "Resource": "*",
-            "Effect": "Allow"
-        }
-    ]
-}
-EOF
-
-}
-
-resource "aws_iam_role_policy_attachment" "titiler_gw_base_policy" {
-  role       = aws_iam_role.titiler_gw_role.name
-  policy_arn = aws_iam_policy.titiler_gw_policy.arn
 }