+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added SSM parameters for Cirrus pre-batch and post-batch lambda function ARNs (`CIRRUS_PRE_BATCH_LAMBDA_ARN` and `CIRRUS_POST_BATCH_LAMBDA_ARN`)
 
+- Added SSM parameter for Cirrus process queue ARN (`CIRRUS_PROCESS_QUEUE_ARN`)
+
 ### Change
 
 ### Fixed